+++ concert/directors/base.py
@@ -3,8 +3,6 @@ import logging
 from abc import abstractmethod
 
 from concert.base import background, Parameter, StateError, Selection, RunnableParameterizable
-from concert.helpers import get_state_from_awaitable
-from concert.loghandler import AsyncLoggingHandlerCloser
 
 LOG = logging.getLogger(__name__)
 
@@ -18,7 +16,8 @@ class Director(RunnableParameterizable):
     current_iteration_name = Parameter()
     log_level = Selection(['critical', 'error', 'warning', 'info', 'debug'])
 
-    async def __ainit__(self, experiment):
+    async def __ainit__(self, experiment, walker=None, separate_scans=True,
+                        name_fmt='director_scan_{:>04}'):
         """
         :param experiment: Experiment that is run. If the experiment features a
             'ready_to_prepare_next_sample' event (asyncio.Event) this will be waited within the
@@ -33,23 +32,16 @@ class Director(RunnableParameterizable):
         self._run_event = asyncio.Event()
         # Let us run by default
         self._run_event.set()
-        self._iteration = 0
-        self.log = LOG
-        self.log.setLevel("INFO")
-        await super().__ainit__()
+        await super().__ainit__(walker=walker, separate_scans=separate_scans,
+                        name_fmt=name_fmt)
+        await self.set_log_file_prefix("director")
 
     async def _get_state(self):
-        state = await get_state_from_awaitable(self._run_awaitable)
+        state = await super().get_state()
         if state == 'running' and not self._run_event.is_set():
             return 'paused'
         else:
             return state
-
-    async def _get_log_level(self):
-        return logging.getLevelName(self.log.getEffectiveLevel()).lower()
-
-    async def _set_log_level(self, level):
-        self.log.setLevel(level.upper())
 
     @abstractmethod
     async def _prepare_run(self, iteration: int):
@@ -92,18 +84,7 @@ class Director(RunnableParameterizable):
     async def _run(self):
         await self._experiment['separate_scans'].stash()
         await self._experiment.set_separate_scans(False)
-        handler = None
         try:
-            if self._experiment.walker:
-                handler: AsyncLoggingHandlerCloser = await self._experiment.walker.register_logger(
-                    logger_name=self.__class__.__name__,
-                    log_level=logging.NOTSET,
-                    file_name="director.log"
-                )
-                self.log.addHandler(handler)
-            self.log.info(await self.info_table)
-
-            await self.prepare()
             # prepare first iteration
             self._iteration = 0
             await self._prepare_run(self._iteration)
@@ -114,13 +95,12 @@ class Director(RunnableParameterizable):
 
                 await self._experiment.walker.descend(await self.get_iteration_name(iteration))
                 exp_run = self._experiment.run()
-                sample_name = await self.get_iteration_name(iteration)
-                self._experiment.log.info(f"Sample name: {sample_name}")
-                self._experiment.log.info(await self.info_table)
+                iteration_name = await self.get_iteration_name(iteration)
+                self._experiment.log.info(f"Iteration name: {iteration_name}")
 
                 # Note: exp_run and self._experiment.ready_to_prepare_next_sample.wait() finish
                 # at the same time, if the user does not implement ready_to_prepare_next_sample.
-                await self._experiment.ready_to_prepare_next_sample.wait()
+                await self._experiment.ready_to_prepare_next_outer_loop.wait()
                 await self._prepare_next_run()
                 try:
                     await exp_run
@@ -144,11 +124,7 @@ class Director(RunnableParameterizable):
             LOG.warning('Experiment director cancelled by keyboard interrupt')
             raise
         finally:
-            if handler:
-                await handler.aclose()
-                self.log.removeHandler(handler)
             await self._experiment['separate_scans'].restore()
-            await self.finish()
 
     async def _get_current_iteration(self) -> int:
         return self._iteration
@@ -167,9 +143,3 @@ class Director(RunnableParameterizable):
         Resumes a currently paused director run.
         """
         self._run_event.set()
-
-    async def prepare(self):
-        pass
-
-    async def finish(self):
-        pass
